backend/requirements.txt
django>=5.1,<5.3
djangorestframework>=3.15,<3.16
django-cors-headers>=4.5,<4.6
python-dotenv>=1.0,<1.1
razorpay>=1.4,<1.5
gunicorn>=23.0,<23.1
uvicorn[standard]>=0.34,<0.35
